requirements.txt
asgiref==3.11.1
astroid==2.15.8
attrs==25.4.0
build==1.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
Django==6.0.2
django-cors-headers==4.9.0
django-split-settings==1.3.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pip-tools==7.5.3
platformdirs==4.9.2
psycopg2-binary==2.9.11
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.11.0
pyproject_hooks==1.2.0
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
ruff==0.15.4
setuptools==82.0.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wheel==0.46.3
wrapt==1.17.3
